# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Incidents, service requests, workflow, assignment, comments and audit for IT service management. Errors are RFC 9457 problem details (`ApiProblem`); clients should branch on `code`.",
    "title" : "Enterprise Service Management API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin as the web app"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "paths" : {
...
    "/api/admin/users" : {
      "get" : {
        "operationId" : "searchUsers",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "role",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          }
        }, {
          "in" : "query",
          "name" : "active",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "array",
            "default" : [ "displayName,ASC" ],
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseUserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Search users by email/name, role and active flag",
        "tags" : [ "Admin: users" ]
      },
      "post" : {
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Create a user with an initial password",
        "tags" : [ "Admin: users" ]
      }
    },
    "/api/admin/users/{id}" : {
      "get" : {
        "operationId" : "getUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Get a user",
        "tags" : [ "Admin: users" ]
      },
      "patch" : {
        "operationId" : "updateUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Rename, change role, activate or deactivate a user (requires current version)",
        "tags" : [ "Admin: users" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiProblem" : {
        "type" : "object",
        "description" : "RFC 9457 problem details, returned for every error",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code: branch on this, never on detail",
            "enum" : [ "VALIDATION_FAILED", "MALFORMED_REQUEST", "AUTHENTICATION_REQUIRED", "ACCESS_DENIED", "RESOURCE_NOT_FOUND", "METHOD_NOT_ALLOWED", "UNSUPPORTED_MEDIA_TYPE", "CONCURRENT_MODIFICATION", "INVALID_SORT_PROPERTY", "INVALID_CREDENTIALS", "REFRESH_TOKEN_INVALID", "INVALID_ORIGIN", "CURRENT_PASSWORD_INCORRECT", "EMAIL_ALREADY_EXISTS", "LAST_ADMIN_REQUIRED", "TEAM_NAME_ALREADY_EXISTS", "TEAM_INACTIVE", "USER_NOT_ELIGIBLE_FOR_TEAM", "CATEGORY_CODE_ALREADY_EXISTS", "CATEGORY_TEAM_REQUIRED", "INVALID_CATEGORY", "INVALID_STATUS_TRANSITION", "TRANSITION_NOT_PERMITTED", "TRANSITION_REQUIREMENT_MISSING", "TICKET_NOT_ASSIGNED", "ASSIGNMENT_NOT_PERMITTED", "ASSIGNEE_NOT_IN_TEAM", "TICKET_NOT_ASSIGNABLE", "MEMBER_HAS_OPEN_TICKETS", "COMMENT_NOT_PERMITTED", "TICKET_CLOSED", "REQUEST_FAILED", "INTERNAL_ERROR" ]
          },
          "correlationId" : {
            "type" : [ "string", "null" ],
            "description" : "Matches the X-Request-Id response header and server log lines"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Human-readable explanation; never contains internal details"
          },
          "fieldErrors" : {
            "type" : [ "array", "null" ],
            "description" : "Present for VALIDATION_FAILED",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorDetail"
            }
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "description" : "Request path",
            "example" : "/api/tickets/0199b2c4-6a1e-7c3e-9f00-1a2b3c4d5e6f/transitions"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 409
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "title" : {
            "type" : "string",
            "example" : "Conflict"
          },
          "type" : {
            "type" : "string",
            "description" : "Problem type URI",
            "example" : "about:blank"
          }
        },
        "required" : [ "code", "correlationId", "detail", "fieldErrors", "instance", "status", "timestamp", "title", "type" ]
      },
...
      "CreateUserRequest" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0
          },
          "initialPassword" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 12
          },
          "role" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          }
        },
        "required" : [ "displayName", "email", "initialPassword", "role" ]
      },
      "FieldErrorDetail" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ]
      },
...
      "PageResponseUserResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "content", "page", "size", "totalElements", "totalPages" ]
      },
...
      "UpdateUserRequest" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0,
            "pattern" : ".*\\S.*"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "version" ]
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "active", "createdAt", "displayName", "email", "id", "role", "updatedAt", "version" ]
      }
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}